Add long-dimension dataset read support
Store dataspace dimensions as long values and add additive long-dimension accessors for datasets and attributes.

Preserve the existing int-based materialization APIs by failing fast for oversized full reads, carry long-dimension handling through contiguous and chunked slice paths, and make implicit chunk indexing lazy for large datasets.

Add regression coverage for oversized dataspaces, contiguous and chunked slice reads, overflow-safe slice validation, and a real large-dimension chunked fixture.

This commit was prepared with OpenAI Codex assistance.

AI-assisted-by: OpenAI Codex

+++ jhdf/src/main/java/io/jhdf/dataset/chunked/indexing/ChunkImpl.java
@@ -21,14 +21,14 @@ public class ChunkImpl implements Chunk {
 
 	private final long address;
 	private final int size;
-	private final int[] chunkOffset;
+	private final long[] chunkOffset;
 	private final BitSet filterMask;
 
-	public ChunkImpl(long address, int size, int[] chunkOffset) {
+	public ChunkImpl(long address, int size, long[] chunkOffset) {
 		this(address, size, chunkOffset, NOT_FILTERED_MASK);
 	}
 
-	public ChunkImpl(long address, int size, int[] chunkOffset, BitSet filterMask) {
+	public ChunkImpl(long address, int size, long[] chunkOffset, BitSet filterMask) {
 		this.address = address;
 		this.size = size;
 		this.chunkOffset = ArrayUtils.clone(chunkOffset);
@@ -46,7 +46,7 @@ public class ChunkImpl implements Chunk {
 	}
 
 	@Override
-	public int[] getChunkOffset() {
+	public long[] getChunkOffset() {
 		return ArrayUtils.clone(chunkOffset);
 	}
